requirements.txt
PyOpenGL
PyOpenGL_accelerate
pyrr
numpy
pygame
